Boton de agregar, y formularioCategoria.jsp para poder colocar toda la info nueva

+++ src/main/java/global/models/Categoria.java
@@ -2,7 +2,7 @@ package global.models;
 
 public class Categoria {
     // Encapsulamiento usando wrappers para permitir null
-    private Integer idCategoria;
+    private Long idCategoria;
     private String nombre;
     private String descripcion;
     private int condicion;
@@ -12,7 +12,7 @@ public class Categoria {
     }
 
     // Constructor con parámetros
-    public Categoria(Integer idCategoria, String nombre, String descripcion, int condicion) {
+    public Categoria(Long idCategoria, String nombre, String descripcion, int condicion) {
         this.idCategoria = idCategoria;
         this.nombre = nombre;
         this.descripcion = descripcion;
@@ -20,11 +20,11 @@ public class Categoria {
     }
 
     // Getters y Setters
-    public Integer getIdCategoria() {
+    public Long getIdCategoria() {
         return idCategoria;
     }
 
-    public void setIdCategoria(Integer idCategoria) {
+    public void setIdCategoria(Long idCategoria) {
         this.idCategoria = idCategoria;
     }
